deploy/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
click==8.1.8
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.6.0
gotrue==2.12.0
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
modal==1.0.2
multidict==6.4.4
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
realtime==2.4.3
rich==14.0.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
synchronicity==0.9.12
toml==0.10.2
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.13.2
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
